Add files via upload
Here is a Kaggle competition where a synthetic but realistic data for migration of birds were given. The task was Exploratory data analysis. Here I have used pandas and Numpy to handle the data and matplotlib and seaborn to visualize the data. Furthemore I have applied Scikit-learn to predict whether a bird will migrate or not with a bunch of features. The dataset is given in the folder.